+++ core/controllers/website.php
@@ -67,7 +67,27 @@ Class Core_Website extends Bancha_Controller
 			{
 				$this->config->prepend_language = $this->lang->current_language;
 			}
+
+			if ($this->config->item('homepage_redirect') == FALSE)
+			{
+				//Homepage served without redirect
+				$segments = explode('/', $home);
+
+				$this->uri->uri_string = trim($home, '/');
+				$this->uri->segments = array($segments);
+				$i = 1;
+				$rsegments = array();
+				foreach ($segments as $segment) {
+					$rsegments[$i] = $segment;
+					$i++;
+				}
+				$this->uri->rsegments = $rsegments;
+				$this->router();return;
+			}
+
+			//Homepage served after a 301 redirect
 			redirect(site_url($home), 'location', 301);
+
 		} else {
 			if (!$this->settings->get('is_installed'))
 			{